adding a test class to test all the controller methods. renaming api directory to controllers. updating deps
